jd_analyzer_backend/requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
# psycopg2-binary==2.9.9
# dj-database-url==2.1.0

# API and serialization
djangorestframework-simplejwt==5.3.0
django-filter==23.3
# Pillow==10.1.0

# AI and ML dependencies
openai==1.3.5
langchain==0.0.340
langchain-openai==0.0.2
tiktoken==0.5.1
# numpy==1.24.3
pandas==2.0.3

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-magic==0.4.27

# Text processing
# nltk==3.8.1
# spacy==3.7.2
textblob==0.17.1

# HTTP and requests
requests==2.31.0
httpx==0.25.2

# Caching and performance
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Monitoring and logging
sentry-sdk==1.38.0
django-extensions==3.2.3

# Security
cryptography==41.0.7
python-decouple==3.8

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development and testing (optional)
django-debug-toolbar==4.2.0
